2006171608 功能修改 修改智能客服功能 移除部分user_con功能 增加部分數據判斷強度 dbdata.sql修改 修改部分dao語法及靜態使用方式 增加繁簡轉換功能SCtable.php sys 修改mysqlDB.php的select方法 修正時區判斷功能 verify.php依照db table名稱修改

+++ controllers/whitelist_con.php
@@ -43,7 +43,7 @@ class whitelist_con
         $ip = $_POST["ip"];
         if (!filter_var($ip, FILTER_VALIDATE_IP)) return false;
         $id = $_POST["id"];
-        if (!is_numeric($id)) return false;
+        if (!is_numeric($id) || empty($id)) return false;
         $wDao = new whitelist_dao;
         return $wDao->setIP($id, $ip);
     }
@@ -52,6 +52,7 @@ class whitelist_con
     {
         if (!key_exists("value", $_POST)) return false;
         $value = $_POST["value"];
+        if (strlen($value) > 1) return false;
         if (!in_array($value, [0, 1])) return false;
         $wsDao = new web_set_dao;
         if (empty($wsDao->getWebSetListBySetKey("whitelist_switch"))) return $wsDao->setWebSetAdd("whitelist_switch", $value);
@@ -62,7 +63,7 @@ class whitelist_con
     {
         if (!key_exists("id", $_POST)) return false;
         $id = $_POST["id"];
-        if (!is_numeric($id)) return false;
+        if (!is_numeric($id) || empty($id)) return false;
         $wDao = new whitelist_dao;
         return $wDao->deleteIP($id);
     }
